requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
biopython
pybedtools
pysam